removed dead code in ResultListenerProtocol
