+++ src/main/scala/edu/ie3/osmogrid/io/output/ResultListener.scala
@@ -6,51 +6,123 @@
 
 package edu.ie3.osmogrid.io.output
 
-import akka.actor.typed.{ActorRef, Behavior, PostStop}
-import akka.actor.typed.scaladsl.Behaviors
+import akka.actor.typed.Behavior
+import akka.actor.typed.ActorRef
+import akka.actor.typed.PostStop
+import akka.actor.typed.scaladsl.{ActorContext, Behaviors, StashBuffer}
 import edu.ie3.datamodel.models.input.container.{
   GridContainer,
   JointGridContainer
 }
 import edu.ie3.osmogrid.cfg.OsmoGridConfig
-import edu.ie3.osmogrid.guardian.OsmoGridGuardian
+import edu.ie3.osmogrid.cfg.OsmoGridConfig.Output
+import edu.ie3.osmogrid.exception.IllegalConfigException
+import edu.ie3.osmogrid.io.output.ResultListenerProtocol._
 
+import concurrent.ExecutionContext.Implicits.global
 import java.util.UUID
+import scala.concurrent.Future
+import scala.util.{Failure, Success}
 
 object ResultListener {
-  // todo JH fully remove
-//  sealed trait Request
-//  final case class GridResult(
-//      grid: GridContainer,
-//      replyTo: ActorRef[Response]
-//  ) extends Request
-//      with ResultEvent
-//  object Terminate extends Request with ResultEvent
-//
-//  sealed trait Response
-//  final case class ResultHandled(
-//      runId: UUID,
-//      replyTo: ActorRef[ResultListener.ResultEvent]
-//  ) extends Response
-//
-//  /* internal API */
-//  sealed trait ResultEvent
-//
-//  def apply(runId: UUID, cfg: OsmoGridConfig.Output): Behavior[ResultEvent] =
-//    Behaviors
-//      .receive[ResultEvent] {
-//        case (ctx, GridResult(grid, replyTo)) =>
-//          ctx.log.info(s"Received grid result for run id '${runId.toString}'")
-//          // TODO: Actual persistence and stuff, ...
-//          replyTo ! ResultHandled(runId, ctx.self)
-//          Behaviors.stopped { () => cleanUp() }
-//        case (ctx, Terminate) => Behaviors.stopped { () => cleanUp() }
-//      }
-//      .receiveSignal { case (ctx, PostStop) =>
-//        ctx.log.info("Requested to stop.")
-//        cleanUp()
-//        Behaviors.same
-//      }
-//
-//  private def cleanUp(): Unit = ???
+
+  def apply(
+      runId: UUID,
+      cfg: OsmoGridConfig.Output
+  ): Behavior[ResultListenerProtocol] = {
+    Behaviors.withStash[ResultListenerProtocol](100) { buffer =>
+      Behaviors.setup[ResultListenerProtocol] { ctx =>
+        ctx.pipeToSelf(initSinks(runId, cfg)) {
+          case Success(sink) =>
+            InitComplete(ListenerStateData(runId, ctx, buffer, sink))
+          case Failure(cause) =>
+            InitFailed(cause)
+        }
+        init(runId, cfg, ctx, buffer)
+      }
+    }
+  }
+
+  private def idle(
+      stateData: ListenerStateData
+  ): Behavior[ResultListenerProtocol] =
+    Behaviors
+      .receiveMessagePartial[ResultListenerProtocol] {
+        case gridResult @ GridResult(grid, replyTo) =>
+          stateData.ctx.pipeToSelf(stateData.sink.handleResult(gridResult)) {
+            case Success(_) =>
+              ResultHandlingSucceeded(
+                ResultHandled(stateData.runId, stateData.ctx.self)
+              )
+            case Failure(exception) =>
+              ResultHandlingFailed(exception)
+          }
+          save(stateData, replyTo)
+      }
+      .receiveSignal { case (context, PostStop) =>
+        if (!stateData.buffer.isEmpty)
+          context.log.warn(
+            s"Stash of ResultListener is not empty! This indicates an invalid system status!"
+          )
+        stateData.sink.close()
+        context.log.info(s"ResultListener stopped!")
+        Behaviors.stopped
+      }
+
+  private def save(
+      stateData: ListenerStateData,
+      replyTo: ActorRef[Response]
+  ): Behavior[ResultListenerProtocol] =
+    Behaviors.receiveMessage {
+      case ResultHandlingFailed(cause) =>
+        stateData.ctx.log.error(
+          s"Error during persistence of grid result. Shutting down!!",
+          cause
+        )
+        Behaviors.stopped
+      case ResultHandlingSucceeded(resultHandled) =>
+        replyTo ! resultHandled
+        Behaviors.stopped
+      case other =>
+        stateData.buffer.stash(other)
+        Behaviors.same
+    }
+
+  private def init(
+      runId: UUID,
+      cfg: OsmoGridConfig.Output,
+      ctx: ActorContext[ResultListenerProtocol],
+      buffer: StashBuffer[ResultListenerProtocol]
+  ): Behavior[ResultListenerProtocol] = {
+    Behaviors.receiveMessage {
+      case InitComplete(stateData) =>
+        stateData.buffer.unstashAll(idle(stateData))
+      case InitFailed(cause) =>
+        ctx.log.error(s"Cannot instantiate $this!", cause)
+        Behaviors.stopped
+      case other =>
+        // stash all other messages for later processing,
+        buffer.stash(other)
+        Behaviors.same
+    }
+  }
+
+  private def initSinks(
+      runId: UUID,
+      cfg: OsmoGridConfig.Output
+  ): Future[ResultSink] = {
+    cfg match {
+      case Output(Some(csv)) =>
+        Future(
+          ResultCsvSink(runId, csv.directory, csv.separator)
+        )
+      case unsupported =>
+        Future.failed(
+          IllegalConfigException(
+            s"Cannot instantiate sink of type '$unsupported'"
+          )
+        )
+    }
+  }
+
 }
